requirements.txt
Django==4.2
djangorestframework==3.14
Pillow==12.2.0
django-cors-headers==4.3
whitenoise==6.12.0
gunicorn==23.0.0